chore(experiments): split ExperimentView/components.tsx by concern
components.tsx had grown to 956 lines bundling three unrelated concerns
behind a name that told readers nothing. Split into focused modules:

- VariantTag.tsx — variant chip used in many places outside the directory
- StatusTag.tsx — experiment status tag, similarly widely used
- LoadingState.tsx — skeleton placeholder
- ExperimentModals.tsx — Pause / Resume / Finish / EditConclusion modals
  plus the shared ConclusionForm (now module-private)
- PageHeader.tsx — PageHeaderCustom plus its internal ResetButton and
  ExperimentDebugToggle helpers (also now module-private)

Three exports with no remaining callers go away with the old file:
EllipsisAnimation, MetricDisplayTrends, MetricDisplayFunnels.

No symbol renames — every importer just swaps the from path. 19 importer
files updated. No behavior change.
