+++ frontend/src/scenes/experiments/ExperimentView/Info.tsx
@@ -19,10 +19,10 @@ import { experimentLogic, previousRefreshAnalytics } from '../experimentLogic'
 import type { ExperimentSceneLogicProps } from '../experimentSceneLogic'
 import { getExperimentStatus, isExperimentPaused } from '../experimentsLogic'
 import { modalsLogic } from '../modalsLogic'
-import { StatusTag } from './components'
 import { ExperimentDuration } from './ExperimentDuration'
 import { ExperimentReloadAction } from './ExperimentReloadAction'
 import { RunningTime } from './RunningTime'
+import { StatusTag } from './StatusTag'
 
 export function Info({ tabId }: Pick<ExperimentSceneLogicProps, 'tabId'>): JSX.Element {
     const {
