--- frontend/src/scenes/experiments/ExperimentView/components.tsx
@@ -1,956 +1,0 @@
-import clsx from 'clsx'
-import { useActions, useValues } from 'kea'
-import { router } from 'kea-router'
-import { useEffect, useState } from 'react'
-import { TextMorph } from 'torph/react'
-
-import {
-    IconArchive,
-    IconCheckCircle,
-    IconCopy,
-    IconEye,
-    IconFlask,
-    IconGlobe,
-    IconList,
-    IconPause,
-    IconPlay,
-    IconPlusSmall,
-    IconRefresh,
-    IconTrash,
-} from '@posthog/icons'
-import {
-    LemonBanner,
-    LemonButton,
-    LemonDialog,
-    LemonDivider,
-    LemonLabel,
-    LemonModal,
-    LemonSelect,
-    LemonSkeleton,
-    LemonSwitch,
-    LemonTag,
-    LemonTextArea,
-    Link,
-    Tooltip,
-} from '@posthog/lemon-ui'
-
-import { useHogfetti } from 'lib/components/Hogfetti/Hogfetti'
-import { InsightLabel } from 'lib/components/InsightLabel'
-import { PropertyFilterButton } from 'lib/components/PropertyFilters/components/PropertyFilterButton'
-import { superpowersLogic } from 'lib/components/Superpowers/superpowersLogic'
-import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
-import { usePageVisibility } from 'lib/hooks/usePageVisibility'
-import { ButtonPrimitive } from 'lib/ui/Button/ButtonPrimitives'
-import { userHasAccess } from 'lib/utils/accessControlUtils'
-import { addProductIntentForCrossSell } from 'lib/utils/product-intents'
-import { organizationLogic } from 'scenes/organizationLogic'
-import { projectLogic } from 'scenes/projectLogic'
-import { sceneLogic } from 'scenes/sceneLogic'
-import { QuickSurveyType } from 'scenes/surveys/quick-create/types'
-import { QuickSurveyModal } from 'scenes/surveys/QuickSurveyModal'
-import { urls } from 'scenes/urls'
-
-import { SceneTitleSection } from '~/layout/scenes/components/SceneTitleSection'
-import { ScenePanel, ScenePanelActionsSection } from '~/layout/scenes/SceneLayout'
-import { groupsModel } from '~/models/groupsModel'
-import { FunnelsQuery, ProductIntentContext, ProductKey, TrendsQuery } from '~/queries/schema/schema-general'
-import {
-    AccessControlLevel,
-    AccessControlResourceType,
-    ActionFilter,
-    AnyPropertyFilter,
-    ExperimentConclusion,
-    ExperimentStatus,
-} from '~/types'
-
-import { CONCLUSION_DISPLAY_CONFIG, EXPERIMENT_VARIANT_MULTIPLE } from '../constants'
-import { CopyExperimentToProjectModal } from '../CopyExperimentToProjectModal'
-import { DuplicateExperimentModal } from '../DuplicateExperimentModal'
-import { canArchiveExperiment, confirmArchiveExperiment, confirmDeleteExperiment } from '../experimentActions'
-import { experimentLogic } from '../experimentLogic'
-import { getExperimentStatusColor, getExperimentStatusLabel, isExperimentPaused } from '../experimentsLogic'
-import { modalsLogic } from '../modalsLogic'
-import { getVariantColor, isLegacyExperiment } from '../utils'
-import { ExperimentSceneMenuBar } from './ExperimentSceneMenuBar'
-
-export function VariantTag({
-    variantKey,
-    fontSize,
-    className,
-}: {
-    variantKey: string
-    fontSize?: number
-    className?: string
-}): JSX.Element {
-    const { experiment } = useValues(experimentLogic)
-
-    if (variantKey === EXPERIMENT_VARIANT_MULTIPLE) {
-        return (
-            <Tooltip title="Some users were exposed to more than one variant. If this percentage is high, there may be an implementation issue causing inconsistent variant assignment.">
-                <LemonTag type="danger">{variantKey}</LemonTag>
-            </Tooltip>
-        )
-    }
-
-    const variantColor = experiment.feature_flag?.filters.multivariate?.variants
-        ? getVariantColor(variantKey, experiment.feature_flag?.filters.multivariate?.variants)
-        : 'var(--text-muted)'
-
-    if (experiment.holdout && variantKey === `holdout-${experiment.holdout_id}`) {
-        return (
-            <span className={clsx('flex items-center min-w-0', className)}>
-                <div
-                    className="w-2 h-2 rounded-full shrink-0"
-                    // eslint-disable-next-line react/forbid-dom-props
-                    style={{
-                        backgroundColor: variantColor,
-                    }}
-                />
-                <LemonTag type="option" className="ml-2">
-                    {experiment.holdout.name}
-                </LemonTag>
-            </span>
-        )
-    }
-
-    return (
-        <span className={clsx('flex items-center min-w-0', className)}>
-            <div
-                className="w-2 h-2 rounded-full shrink-0"
-                // eslint-disable-next-line react/forbid-dom-props
-                style={{ backgroundColor: variantColor }}
-            />
-            <span
-                className="ml-2 text-xs font-semibold truncate text-secondary"
-                // eslint-disable-next-line react/forbid-dom-props
-                style={fontSize ? { fontSize: `${fontSize}px` } : undefined}
-            >
-                {variantKey}
-            </span>
-        </span>
-    )
-}
-
-export function EllipsisAnimation(): JSX.Element {
-    const [ellipsis, setEllipsis] = useState('.')
-    const { isVisible: isPageVisible } = usePageVisibility()
-
-    useEffect(() => {
-        if (!isPageVisible) {
-            return
-        }
-
-        let count = 1
-        let direction = 1
-
-        const interval = setInterval(() => {
-            setEllipsis('.'.repeat(count))
-            count += direction
-
-            if (count === 3 || count === 1) {
-                direction *= -1
-            }
-        }, 300)
-
-        return () => clearInterval(interval)
-    }, [isPageVisible])
-
-    return <TextMorph as="span">{ellipsis}</TextMorph>
-}
-
-export function PageHeaderCustom(): JSX.Element {
-    const {
-        experiment,
-        isExperimentDraft,
-        isExperimentRunning,
-        isExperimentLaunched,
-        isExperimentStopped,
-        isCreatingExperimentDashboard,
-        experimentLoading,
-        launchExperimentLoading,
-    } = useValues(experimentLogic)
-    const {
-        launchExperiment,
-        archiveExperiment,
-        unarchiveExperiment,
-        createExposureCohort,
-        createExperimentDashboard,
-        updateExperiment,
-        setHogfettiTrigger,
-    } = useActions(experimentLogic)
-    const { currentProjectId } = useValues(projectLogic)
-    const { currentOrganization } = useValues(organizationLogic)
-    const hasMultipleProjects = (currentOrganization?.projects?.length ?? 0) > 1
-    const { openFinishExperimentModal, openPauseExperimentModal, openResumeExperimentModal } = useActions(modalsLogic)
-    const [duplicateModalOpen, setDuplicateModalOpen] = useState(false)
-    const [copyToProjectModalOpen, setCopyToProjectModalOpen] = useState(false)
-    const [surveyModalOpen, setSurveyModalOpen] = useState(false)
-    const { newTab } = useActions(sceneLogic)
-    const { trigger, HogfettiComponent } = useHogfetti()
-
-    useOnMountEffect(() => {
-        setHogfettiTrigger(trigger)
-    })
-
-    const exposureCohortId = experiment?.exposure_cohort
-
-    const canEdit = userHasAccess(
-        AccessControlResourceType.Experiment,
-        AccessControlLevel.Editor,
-        experiment.user_access_level
-    )
-    const canArchive = canEdit && canArchiveExperiment(experiment)
-    const canDelete = canEdit
-
-    const handleArchive = (): void => confirmArchiveExperiment(() => archiveExperiment())
-    const handleDelete = (): void =>
-        confirmDeleteExperiment({
-            projectId: currentProjectId,
-            experiment,
-            onDelete: () => router.actions.push(urls.experiments()),
-        })
-
-    return (
-        <>
-            <ExperimentSceneMenuBar />
-            <SceneTitleSection
-                name={experiment?.name}
-                description={null}
-                resourceType={{
-                    type: 'experiment',
-                }}
-                isLoading={experimentLoading}
-                onNameChange={(name) => updateExperiment({ name })}
-                onDescriptionChange={(description) => updateExperiment({ description })}
-                canEdit={canEdit}
-                renameDebounceMs={0}
-                saveOnBlur
-                actions={
-                    <>
-                        {experiment && isExperimentDraft && (
-                            <div className="flex items-center">
-                                <LemonButton
-                                    type="primary"
-                                    data-attr="launch-experiment"
-                                    onClick={() => launchExperiment()}
-                                    loading={launchExperimentLoading}
-                                    size="small"
-                                >
-                                    Launch
-                                </LemonButton>
-                            </div>
-                        )}
-                        {canArchive && (
-                            <LemonButton type="secondary" status="danger" onClick={handleArchive} size="small">
-                                <b>Archive</b>
-                            </LemonButton>
-                        )}
-                        {experiment && isExperimentRunning && !isExperimentStopped && (
-                            <>
-                                <Tooltip title="Conclude this experiment and decide which variant to keep">
-                                    <LemonButton
-                                        type="primary"
-                                        icon={<IconFlask />}
-                                        onClick={() => openFinishExperimentModal()}
-                                        size="small"
-                                    >
-                                        <b>End experiment</b>
-                                    </LemonButton>
-                                </Tooltip>
-                                <FinishExperimentModal />
-                            </>
-                        )}
-                        {experiment && (
-                            <DuplicateExperimentModal
-                                isOpen={duplicateModalOpen}
-                                onClose={() => setDuplicateModalOpen(false)}
-                                experiment={experiment}
-                            />
-                        )}
-                        {experiment && (
-                            <CopyExperimentToProjectModal
-                                isOpen={copyToProjectModalOpen}
-                                onClose={() => setCopyToProjectModalOpen(false)}
-                                experiment={experiment}
-                            />
-                        )}
-                    </>
-                }
-            />
-            <HogfettiComponent />
-
-            {experiment && (
-                <ScenePanel>
-                    <ScenePanelActionsSection>
-                        <ButtonPrimitive menuItem onClick={() => setDuplicateModalOpen(true)}>
-                            <IconCopy />
-                            Duplicate
-                        </ButtonPrimitive>
-
-                        {hasMultipleProjects && (
-                            <ButtonPrimitive
-                                menuItem
-                                onClick={() => setCopyToProjectModalOpen(true)}
-                                disabledReasons={{
-                                    'Copying is not supported for experiments using legacy metrics.':
-                                        isLegacyExperiment(experiment),
-                                }}
-                            >
-                                <IconCopy />
-                                Copy to project
-                            </ButtonPrimitive>
-                        )}
-
-                        {isExperimentLaunched && (
-                            <>
-                                {exposureCohortId ? (
-                                    // TODO: add custom back button to the destination page
-                                    <Link
-                                        to={urls.cohort(exposureCohortId)}
-                                        buttonProps={{
-                                            menuItem: true,
-                                        }}
-                                        data-attr="view-exposure-cohort"
-                                        onClick={() => newTab(urls.cohort(exposureCohortId))}
-                                    >
-                                        <IconEye /> View exposure cohort as new tab
-                                    </Link>
-                                ) : (
-                                    <ButtonPrimitive
-                                        menuItem
-                                        onClick={() => createExposureCohort()}
-                                        data-attr="create-exposure-cohort"
-                                    >
-                                        <IconPlusSmall /> Create exposure cohort
-                                    </ButtonPrimitive>
-                                )}
-                                <ButtonPrimitive
-                                    menuItem
-                                    onClick={() => createExperimentDashboard()}
-                                    disabledReasons={{
-                                        'Creating dashboard...': isCreatingExperimentDashboard,
-                                    }}
-                                >
-                                    <IconPlusSmall /> Create dashboard
-                                </ButtonPrimitive>
-
-                                {experiment.feature_flag && (
-                                    <ButtonPrimitive
-                                        menuItem
-                                        onClick={() => {
-                                            setSurveyModalOpen(true)
-                                            void addProductIntentForCrossSell({
-                                                from: ProductKey.EXPERIMENTS,
-                                                to: ProductKey.SURVEYS,
-                                                intent_context: ProductIntentContext.QUICK_SURVEY_STARTED,
-                                            })
-                                        }}
-                                    >
-                                        <IconPlusSmall /> Create survey
-                                    </ButtonPrimitive>
-                                )}
-
-                                <LemonDivider />
-
-                                {isExperimentRunning &&
-                                    experiment.feature_flag &&
-                                    (isExperimentPaused(experiment) ? (
-                                        <ButtonPrimitive
-                                            menuItem
-                                            data-attr="resume-experiment"
-                                            onClick={() => openResumeExperimentModal()}
-                                        >
-                                            <IconPlay /> Resume experiment
-                                        </ButtonPrimitive>
-                                    ) : (
-                                        <ButtonPrimitive
-                                            variant="danger"
-                                            menuItem
-                                            data-attr="pause-experiment"
-                                            onClick={() => openPauseExperimentModal()}
-                                        >
-                                            <IconPause /> Pause experiment
-                                        </ButtonPrimitive>
-                                    ))}
-
-                                <ResetButton />
-                            </>
-                        )}
-
-                        <LemonDivider />
-
-                        {canArchive && (
-                            <ButtonPrimitive menuItem data-attr="archive-experiment" onClick={handleArchive}>
-                                <IconArchive /> Archive experiment
-                            </ButtonPrimitive>
-                        )}
-                        {canEdit && experiment.archived && (
-                            <ButtonPrimitive
-                                menuItem
-                                data-attr="unarchive-experiment"
-                                onClick={() => unarchiveExperiment()}
-                            >
-                                <IconArchive /> Unarchive experiment
-                            </ButtonPrimitive>
-                        )}
-
-                        {canDelete && (
-                            <ButtonPrimitive
-                                variant="danger"
-                                menuItem
-                                data-attr="delete-experiment"
-                                onClick={handleDelete}
-                            >
-                                <IconTrash /> Delete experiment
-                            </ButtonPrimitive>
-                        )}
-
-                        <PauseExperimentModal />
-                        <ResumeExperimentModal />
-                    </ScenePanelActionsSection>
-                    <ExperimentDebugToggle />
-                </ScenePanel>
-            )}
-            <QuickSurveyModal
-                context={{ type: QuickSurveyType.EXPERIMENT, experiment }}
-                isOpen={surveyModalOpen}
-                onCancel={() => setSurveyModalOpen(false)}
-            />
-        </>
-    )
-}
-
-function ExperimentDebugToggle(): JSX.Element {
-    const { superpowersEnabled } = useValues(superpowersLogic)
-    const { showDebugPanel } = useValues(experimentLogic)
-    const { toggleDebugPanel } = useActions(experimentLogic)
-
-    if (!superpowersEnabled) {
-        return <></>
-    }
-
-    return (
-        <ScenePanelActionsSection>
-            <LemonSwitch
-                className="px-2 py-1"
-                checked={showDebugPanel}
-                onChange={toggleDebugPanel}
-                fullWidth
-                label="Debug panel"
-            />
-        </ScenePanelActionsSection>
-    )
-}
-
-export function ConclusionForm(): JSX.Element {
-    const { experiment } = useValues(experimentLogic)
-    const { setExperiment } = useActions(experimentLogic)
-
-    return (
-        <div className="space-y-4">
-            <div>
-                <LemonLabel>Conclusion</LemonLabel>
-                <LemonSelect
-                    className="w-full"
-                    dropdownMaxContentWidth={true}
-                    value={experiment.conclusion}
-                    options={Object.values(ExperimentConclusion).map((conclusion) => ({
-                        value: conclusion,
-                        label: (
-                            <div className="py-2 px-1">
-                                <div className="font-semibold mb-1.5">
-                                    <div className="font-semibold flex items-center gap-2">
-                                        <div
-                                            className={clsx(
-                                                'w-2 h-2 rounded-full',
-                                                CONCLUSION_DISPLAY_CONFIG[conclusion].color
-                                            )}
-                                        />
-                                        <span>{CONCLUSION_DISPLAY_CONFIG[conclusion].title}</span>
-                                    </div>
-                                </div>
-                                <div className="text-xs text-muted">
-                                    {CONCLUSION_DISPLAY_CONFIG[conclusion].description}
-                                </div>
-                            </div>
-                        ),
-                    }))}
-                    onChange={(value) => {
-                        setExperiment({
-                            conclusion: value || undefined,
-                        })
-                    }}
-                />
-            </div>
-            <div>
-                <LemonLabel>Comment (optional)</LemonLabel>
-                <LemonTextArea
-                    className="w-full border rounded p-2"
-                    minRows={6}
-                    maxLength={400}
-                    placeholder="Optional details about why this conclusion was selected..."
-                    value={experiment.conclusion_comment || ''}
-                    onChange={(value) =>
-                        setExperiment({
-                            conclusion_comment: value,
-                        })
-                    }
-                />
-            </div>
-        </div>
-    )
-}
-
-export function EditConclusionModal(): JSX.Element {
-    const { experiment } = useValues(experimentLogic)
-    const { updateExperiment, restoreUnmodifiedExperiment } = useActions(experimentLogic)
-    const { closeEditConclusionModal } = useActions(modalsLogic)
-    const { isEditConclusionModalOpen } = useValues(modalsLogic)
-
-    return (
-        <LemonModal
-            isOpen={isEditConclusionModalOpen}
-            onClose={closeEditConclusionModal}
-            title="Edit conclusion"
-            width={600}
-            footer={
-                <div className="flex items-center gap-2">
-                    <LemonButton
-                        type="secondary"
-                        onClick={() => {
-                            restoreUnmodifiedExperiment()
-                            closeEditConclusionModal()
-                        }}
-                    >
-                        Cancel
-                    </LemonButton>
-                    <LemonButton
-                        onClick={() => {
-                            updateExperiment({
-                                conclusion: experiment.conclusion,
-                                conclusion_comment: experiment.conclusion_comment,
-                            })
-                            closeEditConclusionModal()
-                        }}
-                        type="primary"
-                        disabledReason={!experiment.conclusion && 'Select a conclusion'}
-                    >
-                        Save
-                    </LemonButton>
-                </div>
-            }
-        >
-            <ConclusionForm />
-        </LemonModal>
-    )
-}
-
-export function PauseExperimentModal(): JSX.Element {
-    const { experiment } = useValues(experimentLogic)
-    const { pauseExperiment } = useActions(experimentLogic)
-    const { closePauseExperimentModal } = useActions(modalsLogic)
-    const { isPauseExperimentModalOpen } = useValues(modalsLogic)
-
-    return (
-        <LemonModal
-            isOpen={isPauseExperimentModalOpen}
-            onClose={closePauseExperimentModal}
-            title="Pause experiment"
-            width={600}
-            footer={
-                <div className="flex items-center gap-2">
-                    <LemonButton type="secondary" onClick={closePauseExperimentModal}>
-                        Cancel
-                    </LemonButton>
-                    <LemonButton
-                        onClick={() => pauseExperiment()}
-                        type="primary"
-                        status="danger"
-                        disabledReason={!experiment.feature_flag && 'No feature flag linked'}
-                    >
-                        Pause experiment
-                    </LemonButton>
-                </div>
-            }
-        >
-            <div className="space-y-4">
-                <div>
-                    Pausing the experiment will <b>disable the feature flag</b>, preventing any users from seeing the
-                    experiment variants. This is useful when you need to quickly stop exposing users to the experiment.
-                </div>
-                <div>The experiment can be resumed at any time. All collected data will be preserved.</div>
-            </div>
-        </LemonModal>
-    )
-}
-
-export function ResumeExperimentModal(): JSX.Element {
-    const { experiment } = useValues(experimentLogic)
-    const { resumeExperiment } = useActions(experimentLogic)
-    const { closeResumeExperimentModal } = useActions(modalsLogic)
-    const { isResumeExperimentModalOpen } = useValues(modalsLogic)
-
-    return (
-        <LemonModal
-            isOpen={isResumeExperimentModalOpen}
-            onClose={closeResumeExperimentModal}
-            title="Resume experiment"
-            width={600}
-            footer={
-                <div className="flex items-center gap-2">
-                    <LemonButton type="secondary" onClick={closeResumeExperimentModal}>
-                        Cancel
-                    </LemonButton>
-                    <LemonButton
-                        onClick={() => resumeExperiment()}
-                        type="primary"
-                        disabledReason={!experiment.feature_flag && 'No feature flag linked'}
-                    >
-                        Resume experiment
-                    </LemonButton>
-                </div>
-            }
-        >
-            <div className="space-y-4">
-                <div>
-                    Resuming the experiment will <b>enable the feature flag</b>, allowing users to see the experiment
-                    variants again. This will continue the experiment from where it was paused.
-                </div>
-                <div>All previously collected data is preserved and new events will be tracked.</div>
-            </div>
-        </LemonModal>
-    )
-}
-
-export function FinishExperimentModal(): JSX.Element {
-    const { experiment, isSingleVariantShipped, shippedVariantKey } = useValues(experimentLogic)
-    const { finishExperiment, endExperimentWithoutShipping, restoreUnmodifiedExperiment } = useActions(experimentLogic)
-    const { closeFinishExperimentModal } = useActions(modalsLogic)
-    const { isFinishExperimentModalOpen } = useValues(modalsLogic)
-    const { aggregationLabel } = useValues(groupsModel)
-
-    const showReleaseModeChoice = useFeatureFlag('EXPERIMENTS_SHIP_VARIANT_RELEASE_MODE', 'test')
-
-    const [selectedVariantKey, setSelectedVariantKey] = useState<string | null>()
-    const [releaseToEveryone, setReleaseToEveryone] = useState<boolean>(false)
-
-    useEffect(() => {
-        if (experiment.parameters?.feature_flag_variants?.length > 1) {
-            // First test variant selected by default
-            setSelectedVariantKey(experiment.parameters.feature_flag_variants[1].key)
-        }
-    }, [
-        experiment.id,
-        experiment.parameters?.feature_flag_variants?.length,
-        experiment.parameters?.feature_flag_variants,
-    ])
-
-    const aggregationTargetName =
-        experiment.filters.aggregation_group_type_index != null
-            ? aggregationLabel(experiment.filters.aggregation_group_type_index).plural
-            : 'users'
-
-    const handleEndExperiment = (): void => {
-        if (isSingleVariantShipped || !selectedVariantKey) {
-            endExperimentWithoutShipping()
-        } else {
-            // Control variant: preserve pre-flag behavior (prepend catch-all release condition).
-            // Test variant: honor the user's radio choice.
-            finishExperiment({
-                selectedVariantKey,
-                releaseToEveryone: showReleaseModeChoice ? releaseToEveryone : true,
-            })
-        }
-    }
-
-    const releaseOptions = [
-        {
-            value: false,
-            icon: <IconList className="text-lg" />,
-            label: 'Roll out to the experiment population',
-            recommended: true,
-            description: `Only ${aggregationTargetName} already in the experiment see the variant. Per-user variant overrides still apply.`,
-        },
-        {
-            value: true,
-            icon: <IconGlobe className="text-lg" />,
-            label: `Roll out to all ${aggregationTargetName}`,
-            recommended: false,
-            description: `All ${aggregationTargetName} see the variant, including those outside the experiment. Per-user variant overrides are bypassed.`,
-        },
-    ] as const
-
-    return (
-        <>
-            <LemonModal
-                isOpen={isFinishExperimentModalOpen}
-                onClose={() => {
-                    restoreUnmodifiedExperiment()
-                    closeFinishExperimentModal()
-                }}
-                width={600}
-                title="End experiment"
-                footer={
-                    <div className="flex items-center gap-2">
-                        <LemonButton
-                            type="secondary"
-                            onClick={() => {
-                                restoreUnmodifiedExperiment()
-                                closeFinishExperimentModal()
-                            }}
-                        >
-                            Cancel
-                        </LemonButton>
-                        <LemonButton
-                            onClick={handleEndExperiment}
-                            type="primary"
-                            disabledReason={!experiment.conclusion && 'Select a conclusion'}
-                        >
-                            End experiment
-                        </LemonButton>
-                    </div>
-                }
-            >
-                <div className="space-y-4">
-                    {isSingleVariantShipped ? (
-                        <div>
-                            <LemonBanner type="info" className="mb-4">
-                                <b>
-                                    <VariantTag variantKey={shippedVariantKey || ''} />
-                                </b>{' '}
-                                is already rolled out to 100% of {aggregationTargetName}. Ending this experiment will
-                                mark it as complete without changing the feature flag.
-                            </LemonBanner>
-                        </div>
-                    ) : (
-                        <>
-                            <div>
-                                <LemonLabel>Variant to keep</LemonLabel>
-                                {!showReleaseModeChoice && (
-                                    <div className="text-sm text-secondary mb-2">
-                                        The selected variant will be rolled out to{' '}
-                                        <b>100% of {aggregationTargetName}</b>.
-                                    </div>
-                                )}
-                                <div className="w-1/2 mt-1">
-                                    <LemonSelect
-                                        className="w-full"
-                                        data-attr="metrics-selector"
-                                        value={selectedVariantKey}
-                                        placeholder="Select a variant"
-                                        onChange={(variantKey) => {
-                                            setSelectedVariantKey(variantKey)
-                                        }}
-                                        allowClear={true}
-                                        options={
-                                            experiment.feature_flag?.filters.multivariate?.variants?.map(({ key }) => ({
-                                                value: key,
-                                                label: (
-                                                    <div className="deprecated-space-x-2 inline-flex">
-                                                        <VariantTag variantKey={key} />
-                                                    </div>
-                                                ),
-                                            })) || []
-                                        }
-                                    />
-                                </div>
-                            </div>
-                            {showReleaseModeChoice && selectedVariantKey && (
-                                <div className="flex flex-col gap-2">
-                                    <LemonLabel>How to release this variant</LemonLabel>
-                                    <div
-                                        className="grid grid-cols-1 md:grid-cols-2 gap-3"
-                                        role="radiogroup"
-                                        aria-label="How to release this variant"
-                                        data-attr="ship-variant-release-mode"
-                                    >
-                                        {releaseOptions.map((option) => {
-                                            const isSelected = releaseToEveryone === option.value
-                                            return (
-                                                <div
-                                                    key={String(option.value)}
-                                                    role="radio"
-                                                    aria-checked={isSelected}
-                                                    tabIndex={0}
-                                                    className={`rounded p-3 cursor-pointer transition-colors ${
-                                                        isSelected
-                                                            ? 'bg-accent-highlight-light border-2 border-accent'
-                                                            : 'border bg-surface-primary border-primary hover:bg-fill-button-tertiary-hover'
-                                                    }`}
-                                                    onClick={() => setReleaseToEveryone(option.value)}
-                                                    onKeyDown={(e) => {
-                                                        if (e.key === 'Enter' || e.key === ' ') {
-                                                            e.preventDefault()
-                                                            setReleaseToEveryone(option.value)
-                                                        }
-                                                    }}
-                                                    data-attr={`ship-variant-release-mode-${
-                                                        option.value ? 'everyone' : 'population'
-                                                    }`}
-                                                >
-                                                    <div className="flex flex-col gap-1">
-                                                        <div className="flex items-center gap-2">
-                                                            {option.icon}
-                                                            <span className="font-medium flex-1">
-                                                                {option.label}
-                                                                {option.recommended && (
-                                                                    <span className="text-secondary text-xs font-normal ml-1">
-                                                                        (recommended)
-                                                                    </span>
-                                                                )}
-                                                            </span>
-                                                            {isSelected && (
-                                                                <IconCheckCircle className="text-accent text-base" />
-                                                            )}
-                                                        </div>
-                                                        <span className="text-xs text-muted">{option.description}</span>
-                                                    </div>
-                                                </div>
-                                            )
-                                        })}
-                                    </div>
-                                </div>
-                            )}
-                        </>
-                    )}
-                    <ConclusionForm />
-                    {!isSingleVariantShipped && (
-                        <LemonBanner type="info" className="mb-4">
-                            For more precise control over your release, adjust the rollout percentage and release
-                            conditions in the{' '}
-                            <Link
-                                target="_blank"
-                                className="font-semibold"
-                                to={experiment.feature_flag ? urls.featureFlag(experiment.feature_flag.id) : undefined}
-                            >
-                                {experiment.feature_flag?.key}
-                            </Link>{' '}
-                            feature flag.
-                        </LemonBanner>
-                    )}
-                </div>
-            </LemonModal>
-        </>
-    )
-}
-
-export const ResetButton = (): JSX.Element => {
-    const { experiment } = useValues(experimentLogic)
-    const { resetRunningExperiment } = useActions(experimentLogic)
-
-    const onClickReset = (): void => {
-        LemonDialog.open({
-            title: 'Reset analysis?',
-            content: (
-                <>
-                    <div className="text-sm text-secondary max-w-md">
-                        <p>
-                            The experiment start and end dates will be reset and the experiment will go back to draft
-                            mode.
-                        </p>
-                        <p>
-                            All events collected thus far will still exist, but won't be applied to the experiment
-                            unless you manually change the start date after launching the experiment again.
-                        </p>
-                        <p>
-                            The <b>feature flag remains untouched</b>, so variants stay visible to users.
-                        </p>
-                    </div>
-                    {experiment.archived && (
-                        <div className="text-sm text-secondary">Resetting will also unarchive the experiment.</div>
-                    )}
-                </>
-            ),
-            primaryButton: {
-                children: 'Confirm',
-                type: 'primary',
-                onClick: resetRunningExperiment,
-                size: 'small',
-            },
-            secondaryButton: {
-                children: 'Cancel',
-                type: 'tertiary',
-                size: 'small',
-            },
-        })
-    }
-
-    return (
-        <ButtonPrimitive variant="danger" menuItem onClick={onClickReset} data-attr="reset-experiment">
-            <IconRefresh /> Reset analysis
-        </ButtonPrimitive>
-    )
-}
-
-export function StatusTag({ status }: { status: ExperimentStatus }): JSX.Element {
-    return (
-        <LemonTag type={getExperimentStatusColor(status)} className="cursor-default">
-            <b className="uppercase">{getExperimentStatusLabel(status)}</b>
-        </LemonTag>
-    )
-}
-
-export function LoadingState(): JSX.Element {
-    return (
-        <div className="deprecated-space-y-4">
-            <LemonSkeleton className="w-1/3 h-4" />
-            <LemonSkeleton />
-            <LemonSkeleton />
-            <LemonSkeleton className="w-2/3 h-4" />
-        </div>
-    )
-}
-
-export function MetricDisplayTrends({ query }: { query: TrendsQuery | undefined }): JSX.Element {
-    const event = query?.series?.[0] as unknown as ActionFilter
-
-    if (!event) {
-        return <></>
-    }
-
-    return (
-        <>
-            <div className="mb-2">
-                <div className="flex mb-1">
-                    <b>
-                        <InsightLabel action={event} showCountedByTag={true} hideIcon showEventName />
-                    </b>
-                </div>
-                <div className="deprecated-space-y-1">
-                    {event.properties?.map((prop: AnyPropertyFilter) => (
-                        <PropertyFilterButton key={prop.key} item={prop} />
-                    ))}
-                </div>
-            </div>
-        </>
-    )
-}
-
-export function MetricDisplayFunnels({ query }: { query: FunnelsQuery }): JSX.Element {
-    return (
-        <>
-            {(query.series || []).map((event: any, idx: number) => (
-                <div key={idx} className="mb-2">
-                    <div className="flex mb-1">
-                        <div
-                            className="shrink-0 w-6 h-6 mr-2 font-bold text-center text-primary-alt border rounded"
-                            // eslint-disable-next-line react/forbid-dom-props
-                            style={{ backgroundColor: 'var(--color-bg-table)' }}
-                        >
-                            {idx + 1}
-                        </div>
-                        <b>
-                            <InsightLabel action={event} hideIcon showEventName />
-                        </b>
-                    </div>
-                    <div className="deprecated-space-y-1">
-                        {event.properties?.map((prop: AnyPropertyFilter) => (
-                            <PropertyFilterButton key={prop.key} item={prop} />
-                        ))}
-                    </div>
-                </div>
-            ))}
-        </>
-    )
-}
