+++ frontend/src/scenes/experiments/legacy/LegacyExperimentView.tsx
@@ -27,9 +27,10 @@ import { Experiment } from '~/types'
 import { experimentLogic } from '../experimentLogic'
 import type { ExperimentSceneLogicProps } from '../experimentSceneLogic'
 import { experimentSceneLogic } from '../experimentSceneLogic'
-import { LoadingState, PageHeaderCustom } from '../ExperimentView/components'
 import { DistributionModal, DistributionTable } from '../ExperimentView/DistributionTable'
 import { ExperimentWarningBanner } from '../ExperimentView/ExperimentWarningBanners'
+import { LoadingState } from '../ExperimentView/LoadingState'
+import { PageHeaderCustom } from '../ExperimentView/PageHeader'
 import { ReleaseConditionsModal, ReleaseConditionsTable } from '../ExperimentView/ReleaseConditionsTable'
 
 const getFirstPrimaryMetric = (experiment: Experiment): ExperimentTrendsQuery | ExperimentFunnelsQuery | null => {
