+++ frontend/src/scenes/experiments/notebook/NotebookExperimentComponent.tsx
@@ -16,8 +16,8 @@ import {
 import { ResultsTag } from '~/scenes/experiments/components/ResultsTag'
 import { experimentLogic } from '~/scenes/experiments/experimentLogic'
 import { getExperimentStatus } from '~/scenes/experiments/experimentsLogic'
-import { StatusTag } from '~/scenes/experiments/ExperimentView/components'
 import { MicroChart } from '~/scenes/experiments/ExperimentView/Exposures'
+import { StatusTag } from '~/scenes/experiments/ExperimentView/StatusTag'
 import { getChanceToWin, isBayesianResult } from '~/scenes/experiments/MetricsView/shared/utils'
 import { isLegacyExperiment } from '~/scenes/experiments/utils'
